Refactor ProfileCell component: replace manual address input with Address component, integrate Customers component, and remove unused Google Maps API logic for improved code clarity and maintainability.
