+++ web/src/components/Profile/Customers/Customers.tsx
@@ -1,0 +1,10 @@
+const Customers = () => {
+  return (
+    <div>
+      <h2>{'Customers'}</h2>
+      <p>{'Find me in ./web/src/components/Customers/Customers.tsx'}</p>
+    </div>
+  )
+}
+
+export default Customers
